histology changes

+++ src/ephys_alignment_gui/custom_atlas.py
@@ -276,4 +276,4 @@ class CustomAtlas(BrainAtlas):
         IMG = sitk.ReadImage(self.atlas_labels_file)
         # Convert sitk to the (ap, ml, dv) np array needed by BrainAtlas
         IMG2 = sitk.DICOMOrient(IMG,self.read_string)
-        self.label = sitk.GetArrayFromImage(IMG2).astype(np.int32)
+        self.label = sitk.GetArrayFromImage(IMG2).astype(np.int16)
